Added signals benchmark
results:
– pseudosignals is plenty fast, and performance is _very_ similar to
std.signals (slightly higher emit() + connect() overhead, but faster
disconnect(); they're within ~20-30% of each other though)
– connect (the slowest individual call) takes ~1.5 µs in the benchmark,
and emit() ~0.5 µs.

This translates to a (theoretical) overhead of ~0.5ms per 1k calls --
so yeah, nothing to worry about.

(We'll probably have on the order of dozens, not hundreds, of signal
calls on the main thread, and our per-frame budget is 15ms).