fixed bowtie bug w/ secondary rendering method
(was drawing "bowties" (ie. bad/broken geometry) when a line doubled
back on itself. This only happens when the corner has been cutoff
(angle < angle_cutoff), and when the angle between the two segments is
> 90 (dot(...) < 0). Our fix is to just duplicate the last points
twice, but in reverse order when this occurs, and the result looks fine
(at the cost of an extra quad per switchback)).