line smoothing (test)
Figured out how to combine edge distance, _and_ uniform edge width
(prev two commits; you might notice that they conflict with each other).

I don't think that we can use flat attributes (thanks dglsl), but we
can get around this by setting the edge factor as both a signed and
unsigned float value in the vertex shader (the signed value gets
interpolated between -x, +x, while the unsigned (abs) value is a
uniform +x in the fragment shader).

Here's how this algorithm works (should definitely add comments, but
w/e):
– width is extended by +N pixels in both directions
– we calculate a thing called the edgeFactor and pass that in along
with the vertex data (in the z component, since we're not using that)
– the edge factor is a signed value (so it gets interpolated), defined
as +edgeFactor for top vertices, or -edgeFactor for bottom vertices
– in the fragment shader, if abs(interpolatedEdgeFactor) > 1.0, the
pixel is inside the edge border (if not alpha = 1.0), and should be a
gradient from the edge border to the edge of the line fragment (and we
can determine how far it is from the edge by also passing in the non
interpolated edge factor)