fixed wraparound bug
(where insert's acquireSegment was grabbing fetchHead before it was
done! This would cause the queue to hang if the producer threads
terminated early, as a bunch of consumer work would get buried in the
middle of the queue after acquire() broke the chain. Insert grabbing an
active, _not-done_ fetch segment and resetting it is also an obvious
bug...).