Wrote internal keybindings conforming to USB-HID so we're not bound to glfw
– mostly did this b/c while glfw's bindings are nice (based on HID, but
adapted so printable keys are matched to ascii and non-printable keys
are outside of it), the keycode range is much larger than necessary and
doesn't pack nicely for the key state struct, and we don't want to have
to use platform-dependent scancode values (I don't even know what the
upper bound for those are!)
– as such, just decided to implement an intermediate, internal key
representation for gsb (usb-hid keycodes), and as a side benefit of
hardcoding the values in a d enum, the keycodes should now be
printable! (ie. writefln("%s", SbKeyCode.KEY_PAGEUP) => "KEY_A", not
"76", or whatever the hell it is in glfw...)
– and it would be nice to be able to add arbitrary native platforms in
the future, which is why I'm going to all the effort of abstracting
things...