added an intermediate mesh + materials format
The general idea is/will be:
– import mesh data + materials from .obj + .mtl / .fbx
– fixup data (eg. generate normals + tangents; cleanup indices, etc --
this mostly applies to .obj files but _may_ apply to .fbx)
– this gets saved into this intermediate format, which is what the sb
scene graph will load

This data will also get cached on disk to improve model load times in
the future (and will get discarded + recalculated if / when the
original source files are updated (hotloading)).

Said format is completely subject to change; goal is to simply build
some intermediary data structures that are compatible w/ both .obj +
.fbx. Processing will hopefully be fast, but will mostly occur on
background threads (load times > 500ms are undesirable, but ok) and can
be optimized later.