added benchmarking per-function
And found something very odd: The _bulk_ of time spent processing
several sample .obj files is spent processing _uvs_ (~6x as long as
vertices or faces).

This is pretty weird b/c parseUv uses almost exactly the same code as
parseVertex (but is slower, somehow?). Maybe this is b/c some models
are using 3-attrib uvs (we discard the .z component), and this is
triggering a different branch path?