GLContext now global
Decided that convenience of having a global context – and lack of
headaches from having to potentially switch between *multiple* proxy
objects with their own states – outweighed the downsides of using a
single thread-shared global proxy object (plus, can use shared static
ctor / dtor, so guaranteed to be initialized + cleaned up correctly –
and it doesn't actually initialize opengl or anything (just a proxy
that records + optimizes state + calls), so this is ok).

This instance is *not* threadsafe, but neither is opengl. If we were
using multiple (real) contexts on different threads, this may cause
issues.